move pages intro draft dir
